Move two scripts from leo/plugins to leo/scripts
